Add files via upload
1. Create a Transaction class include transaction number, date, type, amount. Use type of string for date. There is no need to use Date or DateTime object. For the assignment, you only need, the constructor and toString() methods in the Transaction class. Automatically generate transaction number from a static number and include toString() override for all variables and then create an array to hold the following transactions

1. 16/08/2019     Open Account 100

2. 22/08/2019     Withdraw 50

3. 23/09/2019     Deposit 100
serialize the array and save it to a file named “transactions.bin”read the array from the file and reconstruct a new array from this data and print out each item in the array to the console.

2. Create a BankAccount class with attributes containing name, account number, balance,  transaction and overdraft. Account number should be generated automatically from a static number. You only need the constructor and 4 methods mention below in the BankAccount class.  All variables except overdraft are serializable

Include methods of deposit(String date, double amount), withdraw(String date, double amount), getTransactionDetail(), and toString() override for all variablesprint out “Insufficient fund” message if the withdraw amount is greater than the balance. Transaction with insufficient fund should not be included in the transaction details.

create  a BankAccount object with date “16/08/2017” and amount of 100.  Make withdraw and deposit with following information

22/08/2019 Withdraw 200
23/08/2019 Deposit 100
01/09/2019 Withdraw 50

serialize the BankAccount object and saves it to a file name “accountdetails.bin” and deserialize the file and print out the account details followed by the transaction.

Example of account details:
              Account Number: 1001   Name: Jenny, Lee   Balance: 150.0  

Example of transaction details:
              1. 16/08/2019    Open Account: 100.0
              2. 22/08/2019   Deposit: 100.0
              3. 23/08/2019    Withdraw: 50.0



3. Create a text  file containing the following information without the double quotes
“Would you like to increase your overdraft? Please type Yes/No at the end of the line. ” append “Yes” or “No” at the end of this file using RandomAccessFile object print the appended file to the console. 

The assignment is 9 points + 1 point for the elegance of the code. 
Include your name and student number in a comment in each java source file, and explain your code with comments.